removes the restart handler
The problem with handlers is that they trigger at the end of the play.
This means that if the role is executed for the first time, the service
is first started, as per the final `service: state=started` task, and
then it is restarted because the handler triggers because of the initial
configuration change after the end of the play. This is an unnecessary
restart we can avoid by using the ternary state. This is immediate,
idempotent, and doesn't piss off linting.